requirements.txt
pydantic-settings
python-dotenv
google-genai==1.7.0
streamlit 
requests
streamlit-pdf-viewer==0.0.22
pdfplumber==0.10.0